Fixed created_at and updated_at
Fixed the code in the JSONtoMongo file to utilize the pre 'save'
function to add the created_at and updated_at fields for each listing.
Fixed the code to make sure the updated_at date is saved when Phelps
address is changed. I believe everything is completed for Assignment 4
now.